# code/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Globalstd Challenge API",
        "description": "Documentación de la API de GlobalSTD",
        "contact": {
            "email": "soporte@globalstd.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API GlobalSTD - Servidor principal"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Authenticate a user and return an access token along with user information",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Access token for the authenticated user",
                                            "type": "string"
                                        },
                                        "user": {
                                            "description": "Authenticated user data",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or inactive account"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Invalidate the current access token and logout the user",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/media": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Upload a media file",
                "description": "Store a media file and associate it to the authenticated user",
                "operationId": "uploadMedia",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "collection"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "The file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "collection": {
                                        "description": "The collection or category of the media (e.g., avatar, resume)",
                                        "type": "string",
                                        "example": "avatar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Media file uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "model_type": {
                                            "type": "string",
                                            "example": "App\\\\Models\\\\User"
                                        },
                                        "model_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "collection": {
                                            "type": "string",
                                            "example": "avatar"
                                        },
                                        "disk": {
                                            "type": "string",
                                            "example": "local"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "uploads/avatar123.jpg"
                                        },
                                        "mime_type": {
                                            "type": "string",
                                            "example": "image/jpeg"
                                        },
                                        "original_name": {
                                            "type": "string",
                                            "example": "profile.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}